[console] - added deletion confirmation to various ontology services

+++ console/src/range/Toolbar.tsx
@@ -25,6 +25,7 @@ import { Align } from "@synnaxlabs/pluto/align";
 import { List as Core } from "@synnaxlabs/pluto/list";
 import { Menu as PMenu } from "@synnaxlabs/pluto/menu";
 import { Text } from "@synnaxlabs/pluto/text";
+import { errors } from "@synnaxlabs/x";
 import { useMutation } from "@tanstack/react-query";
 import { type ReactElement, useState } from "react";
 import { useDispatch } from "react-redux";
@@ -32,6 +33,7 @@ import { useDispatch } from "react-redux";
 import { Cluster } from "@/cluster";
 import { ToolbarHeader, ToolbarTitle } from "@/components";
 import { Menu } from "@/components/menu";
+import { Confirm } from "@/confirm";
 import { CSS } from "@/css";
 import { Layout } from "@/layout";
 import { Link } from "@/link";
@@ -43,17 +45,14 @@ import { add, remove, rename, setActive } from "@/range/slice";
 export const List = (): ReactElement => {
   const menuProps = PMenu.useContextMenu();
   const client = Synnax.use();
-  const newLayout = Layout.usePlacer();
+  const placeLayout = Layout.usePlacer();
   const dispatch = useDispatch();
   const ranges = useSelectMultiple();
   const selectedRange = useSelect();
 
   const handleAddOrEdit = (key?: string): void => {
     const layout = createEditLayout(key == null ? "Range.Create" : "Range.Edit");
-    newLayout({
-      ...layout,
-      key: key ?? layout.key,
-    });
+    placeLayout({ ...layout, key: key ?? layout.key });
   };
 
   const handleRemove = (keys: string[]): void => {
@@ -66,25 +65,36 @@ export const List = (): ReactElement => {
 
   const addStatus = Status.useAggregator();
 
+  const confirm = Confirm.useModal();
   const del = useMutation<void, Error, string, Range | undefined>({
-    onMutate: (key: string) => {
+    onMutate: async (key: string) => {
       const rng = ranges.find((r) => r.key === key);
+      if (
+        !(await confirm({
+          message: `Are you sure you want to delete ${rng?.name}?`,
+          description: "This action cannot be undone.",
+          cancel: { label: "Cancel" },
+          confirm: { label: "Delete", variant: "error" },
+        }))
+      )
+        throw errors.CANCELED;
       handleRemove([key]);
       return rng;
     },
     mutationFn: async (key: string) => await client?.ranges.delete(key),
-    onError: ({ message }, _, range) => {
+    onError: (e, _, range) => {
+      if (errors.CANCELED.matches(e)) return;
       addStatus({
         variant: "error",
         message: "Failed to rename range",
-        description: message,
+        description: e.message,
       });
       dispatch(add({ ranges: [range as Range] }));
     },
   });
 
-  const handleSave = (key: string): undefined => {
-    void (async () => {
+  const save = useMutation<void, Error, string, Range | undefined>({
+    mutationFn: async (key: string) => {
       const range = ranges.find((r) => r.key === key);
       if (range == null || range.variant === "dynamic") return;
       await client?.ranges.create({
@@ -93,14 +103,16 @@ export const List = (): ReactElement => {
         name: range.name,
       });
       dispatch(add({ ranges: [{ ...range, persisted: true }] }));
-    })();
-  };
-
-  const handleSetActive = (key: string): void => {
-    void (async () => {
-      await client?.ranges.setActive(key);
-    })();
-  };
+    },
+    onError: (e, _, range) => {
+      addStatus({
+        variant: "error",
+        message: "Failed to save range",
+        description: e.message,
+      });
+      dispatch(add({ ranges: [range as Range] }));
+    },
+  });
 
   const NoRanges = (): ReactElement => {
     const handleLinkClick: React.MouseEventHandler<HTMLParagraphElement> = (e) => {
@@ -133,8 +145,7 @@ export const List = (): ReactElement => {
       edit: () => handleAddOrEdit(rng?.key),
       remove: () => rng != null && handleRemove([rng.key]),
       delete: () => rng != null && del.mutate(rng.key),
-      save: () => rng != null && handleSave(rng.key),
-      setActive: () => rng != null && handleSetActive(rng.key),
+      save: () => rng != null && save.mutate(rng.key),
       link: () => {
         if (rng == null) return;
         const toCopy = `synnax://cluster/${clusterKey}/range/${rng.key}`;
@@ -152,9 +163,6 @@ export const List = (): ReactElement => {
             <Menu.RenameItem />
             <PMenu.Item startIcon={<Icon.Edit />} itemKey="edit">
               Edit
-            </PMenu.Item>
-            <PMenu.Item startIcon={<Icon.Play />} itemKey="setActive">
-              Set as Active Range
             </PMenu.Item>
             <PMenu.Divider />
             <PMenu.Item startIcon={<Icon.Close />} itemKey="remove">
